add superclass for let, ctx and dose plot data models
