+++ pytripgui/canvas_vc/objects/let.py
@@ -1,18 +1,16 @@
 import logging
 
+from pytripgui.canvas_vc.objects.data_base import PlotDataBase
+
 logger = logging.getLogger(__name__)
 
 
-class Let:
+class Let(PlotDataBase):
     def __init__(self, selector):
-        self.aspect = 1.0  # aspect ratio of plot
-
-        self.cube = None
-        self.data_to_plot = None
+        super().__init__(selector)
 
         self.min_let = 0
         self.max_let = None
-        self.projection_selector = selector
 
     def prepare_data_to_plot(self):
         if self.cube is None:
@@ -21,11 +19,3 @@ class Let:
         self._set_aspect()
         self.data_to_plot = self.projection_selector.get_projection(self.cube)
         self.data_to_plot[self.data_to_plot <= self.min_let] = self.min_let
-
-    def _set_aspect(self):
-        if self.projection_selector.plane == "Transversal":
-            self.aspect = self.cube.dimx/self.cube.dimy
-        elif self.projection_selector.plane == "Sagittal":
-            self.aspect = self.cube.dimy*self.cube.pixel_size/(self.cube.dimz*self.cube.slice_distance)
-        elif self.projection_selector.plane == "Coronal":
-            self.aspect = self.cube.dimx*self.cube.pixel_size/(self.cube.dimz*self.cube.slice_distance)
